backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pycryptodome==3.21.0
sqlalchemy==2.0.36
paho-mqtt==2.1.0
websockets==14.1
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
